>  The problems were a bad console log at the top stopping the rest of
stopping the rest of the code from running but i thought it was a bad
first function. I ended up starting over to find the bad bit but nothing
worked.  Second thing was not using let in the loop but that wasnt a big
deal so much. Third was I didnt use === but used = instead.  That was
purely because i wasnt thinking clearly.